💄 design: 추천검색어 모달 구현
