+++ apps/web/src/app/(needLogin)/main/_components/MapPage/_components/Map/Map.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React, { useEffect, useRef, useState } from 'react';
 import useDeepCompareEffect from 'use-deep-compare-effect';
