+++ apps/web/src/app/(needLogin)/main/_components/MapPage/_components/Marker/Marker.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React, { useEffect, useState } from 'react';
 
 export const Marker: React.FC<google.maps.MarkerOptions> = (options) => {
