+++ apps/web/src/app/(needLogin)/main/_components/MapPage/_components/SearchBar/SearchBar.tsx
@@ -1,9 +1,10 @@
-import SearchInput from './SearchInput';
+import { SearchInput, SelectRange } from './_components';
 
 export const SearchBar: React.FC = () => {
   return (
     <div className='absolute left-[35vw] top-8'>
-      <div className='border-primaryGray-100 relative m-auto min-h-[60px] w-[620px] rounded border-[1px]'>
+      <div className='border-primaryGray-100 relative m-auto flex min-h-[50px] w-[580px] rounded border-[1px]'>
+        <SelectRange />
         <SearchInput />
       </div>
     </div>
