+++ apps/web/src/app/(needLogin)/main/_components/MapPage/_components/SearchBar/_components/SearchInput/SearchInput.tsx
@@ -1,12 +1,12 @@
 import { useSetAtom } from 'jotai';
 import { Controller, useForm } from 'react-hook-form';
 
-import { MapPageConfig } from '../../../../../../../constants';
-import { MapAtom } from '../../../../../../../store';
-import materialIcon from '../../../../../../../utils/materialIcon';
-import { SuggestionModal } from './SuggestionModal';
+import { MapPageConfig } from '../../../../../../../../../constants';
+import { MapAtom } from '../../../../../../../../../store';
+import materialIcon from '../../../../../../../../../utils/materialIcon';
+import { SuggestionModal } from '../SuggestionModal';
 
-const SearchInput: React.FC = () => {
+export const SearchInput: React.FC = () => {
   const { handleSubmit, control } = useForm();
   const setIsSearchModalOpen = useSetAtom(MapAtom.isSearchModalOpen);
 
@@ -21,14 +21,14 @@ const SearchInput: React.FC = () => {
           defaultValue=''
           render={({ field }) => (
             <div className='relative mt-1'>
-              <div className='flex min-h-[60px] pt-4'>
+              <div className='flex min-h-[50px] w-[450px]'>
                 <input
                   {...field}
                   onClick={() => setIsSearchModalOpen((prev: boolean) => !prev)}
-                  className='text-align-center text-m ml-12 h-full w-full max-w-[430px] focus:outline-none'
+                  className='text-align-center text-s ml-12 h-full w-[430px] w-full pt-3 focus:outline-none'
                   placeholder={MapPageConfig.searchBarText}
-                ></input>
-                <div className='absolute ml-[380px] cursor-pointer'>
+                />
+                <div className='absolute ml-[380px] cursor-pointer pt-2'>
                   {materialIcon({
                     iconName: 'search',
                     size: 30,
@@ -44,5 +44,3 @@ const SearchInput: React.FC = () => {
     </>
   );
 };
-
-export default SearchInput;
