+++ apps/web/src/app/(needLogin)/main/_components/MapPage/_components/SearchBar/_components/SearchInput/index.ts
@@ -1,0 +1,1 @@
+export * from './SearchInput';
