+++ apps/web/src/app/(needLogin)/main/_components/MapPage/_components/SearchBar/_components/SuggestionModal/SuggestionModal.types.ts
