+++ apps/web/src/app/(needLogin)/main/_components/MapPage/_components/SearchBar/_components/SuggestionModal/index.ts
@@ -1,0 +1,2 @@
+export * from './SuggestionModal';
+export type * from './SuggestionModal.types';
