+++ apps/web/src/app/(needLogin)/main/_components/MapPage/_components/SearchBar/index.ts
@@ -1,2 +1,1 @@
 export * from './SearchBar';
-export * from './SuggestionModal';
